fix(p2): cast EXTRACT(EPOCH) to float8, guard replication on standby, document B4 gap
Final whole-branch review findings:
- numeric->Decimal is not JSON-serializable on PG14+ (blocking, connection_depth, vacuum_horizon)
- pg_current_wal_lsn() errors during recovery, breaking replication.py on standbys
- wait_events' fixed per-target sleep is not counted by the B4 latency-budget check (documented, not rewired)
- add json.dumps() regression guard to the live analyzer pipeline test

+++ .agents/skills/db-report-generator/scripts/collectors/blocking.py
@@ -4,7 +4,7 @@ from scripts.collectors import base
 _SQL = """
 SELECT blocked.pid, blocked.usename, blocking.pid, blocking.usename,
        blocked.query, blocking.query,
-       EXTRACT(EPOCH FROM (now() - blocked.query_start))
+       EXTRACT(EPOCH FROM (now() - blocked.query_start))::float8
 FROM pg_stat_activity blocked
 JOIN LATERAL unnest(pg_blocking_pids(blocked.pid)) AS bp(pid) ON true
 JOIN pg_stat_activity blocking ON blocking.pid = bp.pid
